python/custom-chatGPT-data/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
faiss-cpu==1.7.4
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
numpy==1.25.1
openai==0.27.8
requests==2.31.0
tqdm==4.65.0
urllib3==2.0.3
yarl==1.9.2
